refactor: LottoBundle 클래스 제거
- 제거 이유: WinningLotto 클래스가 당첨 여부 확인 로직을 처리하고, LottoResult에서 당첨 결과를 가공하는 로직을 옮겨서 LottoBundle의 역할이 미비해짐

- LottoResult.java: 등수별로 분리한 리스트를 저장하고 있는 Map을 인자로 받아서 Map<Prize, Integer> 형태로 가공하는 로직을 추가 (countByPrize)

- LottoFactoryTest.java: LottoBundleTest.java에서 수행했던 일부 테스트 코드 이동 (LottoFactory.buyTicket()의 결과 확인)
