+++ src/main/java/lotto/LottoMachine.java
@@ -4,15 +4,17 @@ import lotto.domain.*;
 import lotto.ui.InputView;
 import lotto.ui.ResultView;
 
+import java.util.List;
+
 public class LottoMachine {
 
     public static void main(String[] args) {
-        LottoBundle bundle = LottoFactory.buyTickets(InputView.howMuch());
-        ResultView.showTickets(bundle);
+        List<LottoNumbers> tickets = LottoFactory.buyTickets(InputView.howMuch());
+        ResultView.showTickets(tickets);
 
         WinningLotto winningNumbers = WinningLotto.of(InputView.getWinningNumbers());
 
-        LottoResult result = bundle.drawing(winningNumbers);
+        LottoResult result = winningNumbers.drawing(tickets);
         ResultView.showDrawingResult(result);
     }
 }
